feat: adds configurable JWT settings
Introduces configurable JWT settings, including secret and expiration time,
through the `jwt.config.ts` file and corresponding interface.

This allows for greater flexibility and security by enabling administrators
to easily manage JWT parameters without modifying core application logic.

The changes also include updating the `AuthModule` to read the JWT
configuration from the configuration service, ensuring that the JWT
module is properly configured with the correct secret and expiration time.
A check is added to ensure the JWT configuration is defined and throws an
error if not.

Also, upgrades nx version.